+++ zmu_cortex_m/src/decoder/vldr.rs
@@ -9,7 +9,7 @@ pub fn decode_VLDR_t1(opcode: u32) -> Instruction {
         params: VLoadAndStoreParams {
             dd: ExtensionReg::Double {
                 reg: DoubleReg::from(
-                    opcode.get_bits(12..16) as u8 + ((opcode.get_bit(22) as u8) << 4) as u8,
+                    opcode.get_bits(12..16) as u8 + ((u8::from(opcode.get_bit(22))) << 4),
                 ),
             },
             rn: Reg::from(opcode.get_bits(16..20) as u8),
@@ -26,7 +26,7 @@ pub fn decode_VLDR_t2(opcode: u32) -> Instruction {
         params: VLoadAndStoreParams {
             dd: ExtensionReg::Single {
                 reg: SingleReg::from(
-                    opcode.get_bits(12..16) as u8 + ((opcode.get_bit(22) as u8) << 4) as u8,
+                    opcode.get_bits(12..16) as u8 + ((u8::from(opcode.get_bit(22))) << 4),
                 ),
             },
             rn: Reg::from(opcode.get_bits(16..20) as u8),
